# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentStore API",
    "description": "AI-native digital product store. Purchase BrushFit and other digital products programmatically via x402 protocol, negotiate prices, and submit reviews.",
    "version": "2.0.0",
    "contact": {
      "name": "AgentStore Support",
      "url": "https://agentstore.vercel.app",
      "email": "support@agentstore.vercel.app"
    }
  },
  "servers": [
    {
      "url": "https://agentstore.vercel.app",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/agent/review": {
      "get": {
        "operationId": "getAgentReviews",
        "summary": "Get all agent reviews",
        "description": "Returns all submitted reviews with average rating for BrushFit.",
        "responses": {
          "200": {
            "description": "Reviews list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewList"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "submitAgentReview",
        "summary": "Submit an agent review",
        "description": "Submit a review of BrushFit from an AI agent. Helps build social proof from agent customers.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
    "/api/affiliates": {
      "get": {
        "operationId": "getAffiliateInfo",
        "summary": "Get affiliate program information",
        "description": "Returns details about the BrushFit affiliate program including commission rates, signup instructions, and tracking info.",
        "responses": {
          "200": {
            "description": "Affiliate program details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "program": { "type": "string" },
                    "commission_rate": { "type": "string" },
                    "cookie_duration": { "type": "string" },
                    "signup_url": { "type": "string" },
                    "dashboard_url": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "registerAffiliate",
        "summary": "Register as an affiliate",
        "description": "Sign up for the BrushFit affiliate program to earn commissions on referrals.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "name": { "type": "string" },
                  "website": { "type": "string", "format": "uri" }
                },
                "required": ["email", "name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Affiliate registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "affiliate_id": { "type": "string" },
                    "referral_link": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
  "components": {
    "schemas": {
...
      "ReviewRequest": {
        "type": "object",
        "required": ["agent_id", "rating", "review_text"],
        "properties": {
          "agent_id": { "type": "string", "description": "Unique identifier for the reviewing agent" },
          "rating": { "type": "integer", "minimum": 1, "maximum": 5, "description": "Rating from 1 to 5" },
          "review_text": { "type": "string", "maxLength": 2000, "description": "Review content" },
          "purchase_verified": { "type": "boolean", "default": false, "description": "Whether the agent has verified purchase" }
        }
      },
      "ReviewResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string" },
          "review": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "agent_id": { "type": "string" },
              "rating": { "type": "integer" },
              "review_text": { "type": "string" },
              "purchase_verified": { "type": "boolean" },
              "created_at": { "type": "string", "format": "date-time" }
            }
          }
        }
      },
      "ReviewList": {
        "type": "object",
        "properties": {
          "product": { "type": "string" },
          "total_reviews": { "type": "integer" },
          "average_rating": { "type": "number" },
          "reviews": { "type": "array" }
        }
      }
    }
  }
}